+++ js/toolbarsearchfield.js
@@ -101,6 +101,10 @@
             $(this).removeClass('no-transition');
           });
           this.expand();
+        } else {
+          if (this.button instanceof $ && this.button.length) {
+            this.setClosedWidth();
+          }
         }
 
         return this;
@@ -168,7 +172,6 @@
 
         function searchfieldCollapseTimer() {
           if (!$.contains(self.inputWrapper[0], document.activeElement) && self.inputWrapper.hasClass('active')) {
-            //self.inputWrapper.removeClass('has-focus');
             self.collapse();
           }
         }
@@ -309,9 +312,36 @@
         return distance;
       },
 
+      setClosedWidth: function() {
+        // If the searchfield category button exists, change the width of the
+        // input field on the inside to provide space for the (variable) size of the currently-selected
+        // category (or categories)
+        if ((this.button instanceof $) && this.button.length) {
+          var buttonStyle = window.getComputedStyle(this.button[0]),
+            buttonWidth = parseInt(buttonStyle.width),
+            buttonPadding = parseInt(buttonStyle.paddingLeft) + parseInt(buttonStyle.paddingRight);
+
+          if (this.inputWrapper[0]) {
+            this.inputWrapper[0].style.width = (buttonWidth + buttonPadding) + 'px';
+          }
+        }
+      },
+
       setOpenWidth: function() {
         if (this.inputWrapper[0]) {
           this.inputWrapper[0].style.width = this.openWidth;
+        }
+
+        // If the searchfield category button exists, change the width of the
+        // input field on the inside to provide space for the (variable) size of the currently-selected
+        // category (or categories)
+        if ((this.button instanceof $) && this.button.length) {
+          var buttonStyle = window.getComputedStyle(this.button[0]),
+            buttonWidth = parseInt(buttonStyle.width),
+            buttonPadding = parseInt(buttonStyle.paddingLeft) + parseInt(buttonStyle.paddingRight),
+            inputWidth = 'calc(100% - ' + (buttonWidth + buttonPadding) + 'px)';
+
+          this.input[0].style.width = inputWidth;
         }
       },
 
@@ -458,7 +488,11 @@
 
         function closeWidth() {
           if (self.settings.collapsible || self.shouldBeFullWidth()) {
-            self.inputWrapper.removeAttr('style');
+            if (self.button instanceof $ && self.button.length) {
+              self.setClosedWidth();
+            } else {
+              self.inputWrapper.removeAttr('style');
+            }
           }
         }
 
@@ -477,7 +511,7 @@
 
           closeWidth();
 
-          if (self.button && self.button.length && self.button.is('.is-open')) {
+          if (self.button && self.button.length) {
             self.button.data('popupmenu').close(false, true);
           }